Upgrade to fs-capacitor v4 beta
I've done a rework of fs-capacitor to support node 13. This involved
using the `readable-stream` package instead of node's stream library
in order to simply continued compatibility with node v8.x, which we can
drop at the end of 2019.

I spent the better part of a day trying to figure out exporting
interoperable es6 and commonjs modules now that I have a commonjs
dependency, and I finally just gave up. Without any useful guidance or
pathway put forward by node for interoperability, the only choice here
is to chose a single module system, which will be commonjs for the
immediate future.

I'm very unhappy with this situation and fear this is going to severly
fragment the ecosystem, but don't have the time to get involved any
further.